Edured-79: Moved ErrorData to it's own file in the main package
