+++ app/src/main/kotlin/nl/eduid/ErrorData.kt
@@ -1,0 +1,3 @@
+package nl.eduid
+
+data class ErrorData(val title: String, val message: String)
